+++ lib/src/cli/commands/commands.dart
@@ -1,3 +1,4 @@
+export 'cache_command.dart';
 export 'config_command.dart';
 export 'crawl_command.dart';
 export 'cues_command.dart';