+++ lib/src/misc/misc.dart
@@ -1,5 +1,7 @@
+export 'config_initializer.dart';
 export 'constants.dart';
 export 'cue_kinds.dart';
+export 'default_configs.dart';
 export 'encoders.dart';
 export 'env.dart';
 export 'fuzzy_search.dart';